Temp fix #389 until 2.18 changelog goes live
The broken link to the visual changelog on the homepage has been reported here ( #389 ) and elsewhere (http://osgeo-org.1560.x6.nabble.com/Broken-Link-to-2-18-Visual-Change-Log-td5292394.html). While the long term fix is, of course, completing the changelog, for the time being this, or something like it seems clearer and less confusing from a user's perspective rather than sending them to a 404. The wording can be changed as seen fit. Once the changelog goes live, this change would then be rolled back and all would be well. Thanks for your consideration.