Update commercial_support.rst
Hey - Good morning. I'm updating the NRGS "About" section  (lines 424-427) as I've not revisited in this in several years. I also have a new logo - but I'm not sure what to go about that. Can I email it to someone or add it somewhere? Thanks in advance for any help and just in general - thanks.