Update commercial_support.rst (#845)
* Update commercial_support.rst

Hey - Good morning. I'm updating the NRGS "About" section  (lines 424-427) as I've not revisited in this in several years. I also have a new logo - but I'm not sure what to go about that. Can I email it to someone or add it somewhere? Thanks in advance for any help and just in general - thanks.

* Remving logo

I need to remove this logo and I'm uploading one with the exact same name because.....that sounds like the most reasonable thing to not break anything......

* Replacing NRGS Logo

Replacing the recently removed NRGS logo with this one - which should be the same name and size.

* Update source/site/forusers/commercial_support.rst

Co-authored-by: Harrissou Sant-anna <delazj@gmail.com>