Additions for Fedora, Opensuse and general Linux
I started by adding up-to-date version infos for Fedora. Then I added the instructions for Toolbox and Distrobox, toolbox being standard on Fedora Atomic, Distrobox on Opensuse microOS but also working on Fedora Atomic. 

These tools create containers of a Distro, so the QGis packages can be isolated from the host system.

GUI app entries are included in the explanations.

---

Then I added an instruction to add missing python dependencies using pip, under the main Linux chapter.

I also added a tutorial how to enforce XWayland, as Wayland shows a warning of missing features. This is opt-in but automated.